Potions are now only deleted when Brew rewards are closed by user
Fixes #3

Previously, if a user opened e.g. the map or the deck while the brew
rewards were open, all potions would be deleted due to the way closing
the screen was detected. Now the closing detected is a user actually
clicking the close button and not just switching to another screen